Game over screen
Game over widget displays with option of trying again or giving up. If the player choses to try again, they will return to their spot in the over world as if they hadn't started the battle yet. The "give up" button does not do anything yet; I assume we want to add a title screen that we could return this back to.